add javascript dom manipulation methods
Group all dom manipulation methods in a section.

Add the code to extend the chat slider. It inspects the
slider height to determine if it's fully retracted. If so,
it uses a jQuery animation to extend it.

Add the code to retract the chat slider. It inspects the
slider height to determine if it's fully extended. If so,
it uses a jQuery animation to retract it.