setup and export our only public method
Add code to fill the $container with the slider template html.
Find the chat slider div and store it in a module-scope variable,
$chatSlider. A module-scope variable is available to all functions
in the spa namespace.

Set the title of the slider and bind the onClickSlider handler to
a click event on the chat slider.

Export public methods by returning an object from our spa namespace.
We export only one method - initModule.